+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Split files by 1 MB by default
+
 ## [0.4.7] - 2026-04-14
 
 ### Added